added common theme to avoid duplication
